+++ lib/src/analyzers/unused_localization_analyzer/models/unused_localization_issue.dart
@@ -1,0 +1,1 @@
+import 'package:source_span/source_span.dart';